+++ assets/javascripts/kitten/components/buttons/button/test.js
@@ -58,6 +58,16 @@ describe('<Button />', () => {
     })
   })
 
+  describe('with `radius` prop', () => {
+    beforeEach(() => {
+      component = renderer.create(<Button radius={4}>MyButton</Button>).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
   describe('with `as` prop', () => {
     beforeEach(() => {
       component = renderer